Replace DirectoryWatcher with a more generic version.
I have copied the DirectoryWatcher into my MiscUtils package. To do that,
I had to remove the step where it converted the WatcherChangeTypes. Instead,
that is now done in Lithogen.exe.

The reason for this change is that I want a more generic DirectoryWatcher
so that I can use it to implement the LiveReloadServer.