Get LiveReload server side working, but it's crude at the moment.
We need the EmbedIO upstream to make some changes to their server to
allow us to inject the directory into the constructor, and to fix their
IDisposable implementation.

Do we also need to consider state management? The LiveReload script
disconnects and reconnects after each refresh of the main page on the GCS.